#110: Extract the DatabaseManager interface
